Version 1.44 (2008-02-29)
As of version 1.44, the command-line options and the configuration of the
qlicserver have changed. The goal of the changes is to improve the ease of
deployment and to increase the flexibility of the program in general.
Editing the source code should no longer be required, but can be done if you
either wish to operate without a configuration file, or would like to have
some failsafe values hard-coded in. Although upgrading from a previous
version should be straightforward, please do check if there is now a
different way of doing things or if a particular option has been eliminated.
The -h option usually provides a good starting point.

The most notable changes:

    * XML configuration file

A configuration file can be used for most of the license configuration and
operation parameters -- qlicserver.config

    * Resource limit files

Resource limits controllable from a file or a directory of files --
qlicserver.limits

    * XML output format

With supplementary XSLT sheets it allows web-based monitoring of the
resources.

    * Optional qhost/qstat output

These results of these queries can be used by processed by appropriate XSLT
sheets (eg, within xmlqstat) to monitor the status of jobs and hosts.

    * Silent operation by default

The previous -q (quiet) option has been dropped in favour of explicitly
requesting output with the output parameter.

    * Command-line parameters

Many program parameters can also be set in the configuration file or via the
command-line.

    * More flexible daemon modes

It is now possible to operate like a true daemon (without a parent) or to
monitor a particular process or process id.