Release 2009-04-28
- fixed regression occurring with derived resources

  missed one level of indirection when combining derived resources

    *  version numbering

The qlicserver version numbers have been dropped in favour of simply using
the release date. This is partially convenience and partially a side-effect
of changing from cvs to git control.

    * GridEngine 'JOB' complex

Complexes with the type 'job' are not scaled by the number of slots.

    * bugfix for derived resource

Derived resources should now correctly respect the limits imposes on their
subelements.

    * initial support for license server triads

Elements in the LM_LICENSE_FILE surrounded by brace brackets are taken
literally and not threaded. This helps prevent querying a license triad
three times and getting the wrong answer. Using brace brackets to group and
identify triads is probably only meaningful within a config file and should
not be attempted on the LM_LICENSE_FILE environment variable itself.

    * license remapping

With the from attribute, it is possible to map FlexLM licenses to different
GridEngine resource names depending on the license server. This can be used,
for example, to distinguish between identically named license features that
have different qualities (eg, an old purchase license vs. a new lease
license).

    * atomic write operation

Output files are written to a temporary file and unlink() is used to move
them to the correct location. This avoids race conditions when the file is
already being read by another process.